+++ types.d.ts
@@ -19,13 +19,13 @@ declare namespace globalThis {
 		readonly overrides?: {
 			readonly general?: {
 				readonly 'no-restricted-globals'?: {
-					readonly additionalRestictions?: string[];
+					readonly additionalRestrictions?: string[];
 				};
 				readonly 'no-restricted-properties'?: {
-					readonly additionalRestictions?: string[];
+					readonly additionalRestrictions?: string[];
 				};
 				readonly 'no-restricted-syntax'?: {
-					readonly additionalRestictions?: string[];
+					readonly additionalRestrictions?: string[];
 				};
 				readonly 'import/order'?: {
 					readonly additionalExternalPatterns?: string[];